+++ castor/src-js-native/ContextCompanionObject.scala
@@ -1,0 +1,3 @@
+package castor.platform
+
+private [castor] trait ContextCompanionObject
