+++ castor/src-js-native/ContextImpl.scala
@@ -1,0 +1,3 @@
+package castor.platform
+
+private [castor] trait ContextImpl
