+++ castor/src-js-native/ContextSimpleCompanionObject.scala
@@ -1,0 +1,3 @@
+package castor.platform
+
+private [castor] trait ContextSimpleCompanionObject
