+++ castor/src-js-native/Platform.scala
@@ -1,0 +1,7 @@
+package castor.platform
+
+import scala.concurrent.ExecutionContext
+
+private [castor] object Platform {
+  def executionContext = ExecutionContext.global
+}
