+++ castor/src-jvm/castor/platform/Platform.scala
@@ -1,0 +1,11 @@
+package castor.platform
+
+import castor._
+
+import scala.concurrent.ExecutionContext
+
+import java.util.concurrent.{Executors, ThreadFactory, TimeUnit}
+
+private [castor] object Platform {
+  def executionContext = ExecutionContext.fromExecutorService(Context.Simple.threadPool)
+}
